+++ src/pages/main/RequestListPage/index.module.css
@@ -1,0 +1,59 @@
+.request_entry {
+  padding: 16px;
+  border: 1px solid var(--itmo-colors-gray);
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.request_header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 8px;
+}
+
+.request_titles {
+  font-weight: bold;
+  color: var(--itmo-colors-black);
+}
+
+.request_text {
+  color: #333333;
+  overflow-wrap: break-word; /* Ensures the break of long words if necessary */
+  word-wrap: break-word;     /* Deprecated version of the above property for older browsers */
+  white-space: normal;       /* Allows the text to wrap to the next line */
+}
+
+.request_respond {
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 8px;
+}
+
+.approve_button {
+  color: white;
+  border: none;
+  padding: 8px 16px;
+  background-color: var(--itmo-colors-gray);
+}
+
+.approve_button:hover {
+  cursor: auto;
+  background-color: var(--itmo-colors-gray);
+}
+
+.decline_button {
+  color: white;
+  border: none;
+  padding: 8px 16px;
+  background-color: var(--itmo-colors-gray);
+}
+
+.decline_button:hover {
+  cursor: auto;
+  background-color: var(--itmo-colors-gray);
+}
+
+
